refactor(transaction_factory): move transaction_factory.rs to data_structures
